infra: create Dockerfile to build app in container
Create Dockerfile describing a simple application infraestructure and
create a build process to run chatbot as a self-containered application.

Also create a .dockerignore to ignore `.env` and `node_modules/` because
de environment variables must be provided when the container will be
deployed and the node modules will be installed on the container build
process.