infra: describe development infra using docker-compose
Describe a development infraestructure using docker-compose file. In
this docker-compose.yml file have services:

- Node 8.9
- Nginx

Also create a command to watch modifications on directory that contain
source files using nodemon, that is, every change will be updated, so
its not necessary to restart container.

It was necessary to create nginx configuration file, so this commit also
add a configuration file to nginx. This file just redirect http requests
to node container on port 3000.