bot: define a storage method
The bot persist some data about users. This data is responsible for
identify the user and is used to define who will receive an asware, for
example, but default storage method was deprecated, so bot stop to answare.

To solve this problem, this commit set inMemoryStorage as a storage method.